feat: adicao de novos itens nas classes ja existentes referentes a doacao de alimentos, para poder gerenciar alimentos e gerar uma cesta

+++ projeto/backend/src/main/java/br/com/faitec/foodcare/implementation/service/product/ProductServiceImpl.java
@@ -82,7 +82,7 @@ public class ProductServiceImpl implements ProductService {
     }
 
     @Override
-    public boolean updateQuantity(int id, int newStock) {
+    public boolean updateQuantity(int id, double newStock) {
         Product product = findById(id);
 
         if(product == null){
