+++ projeto/backend/src/main/java/br/com/faitec/foodcare/port/dao/product/UpdateQuantityDao.java
@@ -1,5 +1,5 @@
 package br.com.faitec.foodcare.port.dao.product;
 
 public interface UpdateQuantityDao {
-    boolean updateQuantity(final int id, final int newStock);
+    boolean updateQuantity(final int id, final double newStock);
 }
