+++ projeto/backend/src/main/java/br/com/faitec/foodcare/port/service/product/UpdateQuantityService.java
@@ -1,5 +1,5 @@
 package br.com.faitec.foodcare.port.service.product;
 
 public interface UpdateQuantityService {
-    boolean updateQuantity(final int id, final int newStock);
+    boolean updateQuantity(final int id, final double newStock);
 }
